separated out user settings component
