+++ frontend/src/components/trip-panel/TripPanelControl.js
@@ -1,93 +1,10 @@
-import * as React from "react";
-import {
-	Box,
-	Menu,
-	MenuItem,
-	ListItemIcon,
-	Tooltip,
-	Button,
-	Typography,
-} from "@mui/material";
-import PersonAdd from "@mui/icons-material/PersonAdd";
-import Settings from "@mui/icons-material/Settings";
-import Logout from "@mui/icons-material/Logout";
+import { Box } from "@mui/material";
+import { UserSettings } from "./UserSettings";
 
 export const TripPanelControl = () => {
-	const [anchorEl, setAnchorEl] = React.useState(null);
-	const open = Boolean(anchorEl);
-	const handleClick = (event) => {
-		setAnchorEl(event.currentTarget);
-	};
-	const handleClose = () => {
-		setAnchorEl(null);
-	};
 	return (
 		<Box sx={{ mb: 3 }}>
-			<Button
-				onClick={handleClick}
-				variant="text"
-				sx={{
-					width: "100%",
-					color: "black",
-					justifyContent: "flex-start",
-				}}
-			>
-				<Typography variant="h5">Username</Typography>
-			</Button>
-			<Menu
-				anchorEl={anchorEl}
-				open={open}
-				onClose={handleClose}
-				onClick={handleClose}
-				PaperProps={{
-					elevation: 0,
-					sx: {
-						minWidth: 300,
-						overflow: "visible",
-						filter: "drop-shadow(0px 2px 8px rgba(0,0,0,0.32))",
-						mt: 1.5,
-						"& .MuiAvatar-root": {
-							width: 32,
-							height: 32,
-							ml: -0.5,
-							mr: 1,
-						},
-						"&:before": {
-							content: '""',
-							display: "block",
-							position: "absolute",
-							top: 0,
-							left: 10,
-							width: 10,
-							height: 10,
-							bgcolor: "background.paper",
-							transform: "translateY(-50%) rotate(45deg)",
-							zIndex: 0,
-						},
-					},
-				}}
-				transformOrigin={{ horizontal: "right", vertical: "top" }}
-				anchorOrigin={{ horizontal: "left", vertical: "bottom" }}
-			>
-				<MenuItem>
-					<ListItemIcon>
-						<PersonAdd fontSize="small" />
-					</ListItemIcon>
-					Add another account
-				</MenuItem>
-				<MenuItem>
-					<ListItemIcon>
-						<Settings fontSize="small" />
-					</ListItemIcon>
-					Settings
-				</MenuItem>
-				<MenuItem>
-					<ListItemIcon>
-						<Logout fontSize="small" />
-					</ListItemIcon>
-					Logout
-				</MenuItem>
-			</Menu>
+			<UserSettings />
 		</Box>
 	);
 };
